+++ src/index.ts
@@ -1,12 +1,17 @@
-import { regex } from "./pattern";
+import { fullPattern, regularExpression } from "./pattern";
 import patternsList from "./patterns.json";
+
+/**
+ * Naive bot pattern.
+ */
+const naivePattern = /bot|spider|crawl|http|lighthouse/i;
 
 // Workaround for TypeScript's type definition of imported variables and JSON files.
 
 /**
  * A pattern that matches bot identifiers in user agent strings.
  */
-export const pattern: RegExp = regex;
+export const pattern = regularExpression;
 
 /**
  * A list of bot identifiers to be used in a regular expression against user agent strings.
@@ -14,10 +19,26 @@ export const pattern: RegExp = regex;
 export const list: string[] = patternsList;
 
 /**
+ * Check if the given user agent includes a bot pattern. Naive implementation (less accurate).
+ */
+export const isbotNaive = (userAgent?: string | null): boolean =>
+	Boolean(userAgent) && naivePattern.test(userAgent);
+
+let usedPattern: RegExp;
+/**
  * Check if the given user agent includes a bot pattern.
  */
-export const isbot = (userAgent?: string | null): boolean =>
-	Boolean(userAgent) && pattern.test(userAgent);
+export function isbot(userAgent?: string | null): boolean {
+	if (typeof usedPattern === "undefined") {
+		try {
+			// Build this RegExp dynamically to avoid syntax errors in older engines.
+			usedPattern = new RegExp(fullPattern, "i");
+		} catch (error) {
+			usedPattern = naivePattern;
+		}
+	}
+	return Boolean(userAgent) && usedPattern.test(userAgent);
+}
 
 /**
  * Create a custom isbot function with a custom pattern.