+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [4.4.0](https://github.com/omrilotan/isbot/compare/v4.3.0...v4.4.0)
+
+- Add a naive fallback pattern for engines that do not support lookbehind in regular expressions
+- Add isbotNaive function to identify bots using a naive approach (simpler and faster)
+
 ## [4.3.0](https://github.com/omrilotan/isbot/compare/v4.2.0...v4.3.0)
 
 - Accept `undefined` in place of user agent string to allow headers property to be used "as is" (`request.headers["user-agent"]`)